NBayes {NBayes}	R Documentation

Naive Bayes Classifier

Description

This is the fit function of the classifier using binarization on categorical predictors in order to project them in a gaussian continous space.

Usage


NBayes(formula, data)

Arguments

formula
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.

data
matrix or dataframe with categorical (type of character or factor or logical) or numeric predictors and the categorical dependant variable.

Value

a S3 class object "NBayes"

Note

For prediction purpose, please input the S3 class object "NBayes" in the generic function predict()

Author(s)

Long Nguyen Phuoc <long.nguyenph@gmail.com>

See Also

predict.NBayes

Examples


NBayes(Species ~ .,iris)
NBayes(Species ~ Sepal.Length + Sepal.Width,iris)

[Package NBayes version 0.0.0.9000 Index]