predict.NBayes {NBayes}	R Documentation

Predict Method for NBayes Objects

Description

This is a method for the predict generic function on object NBayes inheriting from NBayes.

Usage


## S3 method for class 'NBayes'
predict(object, newdata, type = c("class", "posterior"))

Arguments

object
an object of class "NBayes" inheriting from NBayes.

newdata
matrix or dataframe with categorical (type of character or factor or logical) or numeric predictors and without the categorical dependant variable.

type
if "class", vector of dependent variables predicted and classified according to the highest posterior probabilities. If "posterior", the posterior probabilities for each modality are returned in a matrix.

Value

if type="class", vector of dependent variables predicted and classified according to the highest posterior probabilities. If type="posterior", the posterior probabilities for each modality are returned in a matrix.

Author(s)

Long Nguyen Phuoc <long.nguyenph@gmail.com>

See Also

NBayes()

Examples


model_iris <- NBayes(Species~.,iris)
predict(model_iris,iris[,-5])
predict(model_iris,iris[,-5],type="posterior")

[Package NBayes version 0.0.0.9000 Index]